# feat: implement fast path optimization for event bus (Phase 3)